Support multiple paths in file basedir
The "basedir" keyword arg passed to ContextDict can now have
multiple paths delimited by ":".  When rendering an ContextVal
with a multi-path basedir, the first path to match an existing
file is returned.  If no matches are found then the last path
is returned.  The case of only one path always falls through
to the "last path" so that the ContextValue behavior is unchanged.

Also

- Update version to 0.2.1
- Fix some docs regressions due to the previous commit.

Note:
Code will likely fail on Windows since basedir paths are split
on ":".